AA_Energies/requirements.txt
##############################
# 🧠 CORE Y UTILIDADES
##############################
certifi==2025.10.5              # Certificados SSL/TLS para peticiones seguras
charset-normalizer==3.4.4       # Normalización de texto y codificaciones
idna==3.11                      # Manejo de nombres de dominio internacionales
requests==2.32.5                # Peticiones HTTP
urllib3==2.5.0                  # Cliente HTTP usado por requests
python-dateutil==2.9.0.post0    # Manejo avanzado de fechas
pytz==2025.2                    # Soporte para zonas horarias
tzdata==2025.2                  # Datos de zonas horarias para sistemas sin tzdata

##############################
# 🧱 DJANGO Y FRAMEWORKS WEB
##############################
Django==5.2.7                   # Framework principal
djangorestframework==3.16.1     # API REST con Django
django-import-export==4.3.13    # Importar/exportar datos (CSV, XLSX, JSON, etc.)

##############################
# 🐘 BASE DE DATOS
##############################
psycopg2-binary==2.9.11         # Driver PostgreSQL
odfpy==1.4.1                    # Soporte para archivos ODF
openpyxl==3.1.5                 # Lectura/escritura de archivos Excel (.xlsx)
xlrd==2.0.2                     # Lectura de archivos Excel (.xls)
xlwt==1.3.0                     # Escritura de archivos Excel antiguos (.xls)
tablib==3.9.0                   # Biblioteca de manipulación de tablas/datos
et_xmlfile==2.0.0               # XML backend para openpyxl

##############################
# 📊 ANÁLISIS DE DATOS
##############################
numpy==2.3.4                    # Cálculos numéricos
pandas==2.3.3                   # Manipulación de datos/tablas

##############################
# 🧪 TESTING
##############################
pytest==8.4.2                   # Framework principal de testing
pytest-django==4.11.1           # Integración de Django con pytest
pytest-cov==7.11.0              # Cobertura de código en pytest
coverage==7.11.0                # Medición de cobertura
pytest-playwright==0.7.1        # Testing e2e con Playwright
pytest-base-url==2.1.0          # Configurar URL base para tests

##############################
# 🌐 PLAYWRIGHT (AUTOMATIZACIÓN WEB)
##############################
playwright==1.55.0              # Navegación y testing automatizado
pyee==13.0.0                    # Manejador de eventos usado por Playwright

##############################
# 🧹 CALIDAD DE CÓDIGO Y SEGURIDAD
##############################
bandit==1.8.6                   # Análisis de seguridad en código Python
flake8==7.3.0                   # Linter de código
pycodestyle==2.14.0             # Estilo de código (PEP8)
pyflakes==3.4.0                 # Detección de errores en código
black==25.9.0                   # Formateador automático


##############################
# 🧩 UTILIDADES Y FORMATO
##############################
markdown-it-py==4.0.0           # Parser de Markdown
mdurl==0.1.2                    # Manejo de URLs en Markdown
rich==14.2.0                    # Colores y formato avanzado en consola
tabulate==0.9.0                 # Tablas de texto en consola
text-unidecode==1.3             # Normalización de texto Unicode
python-slugify==8.0.4           # Generación de slugs amigables (URLs)
beautifulsoup4==4.14.2          # Análisis de HTML/XML
soupsieve==2.8                  # Selectores CSS para BeautifulSoup
defusedxml==0.7.1               # Seguridad para XML
diff-match-patch==20241021      # Comparación de texto (diffs)
stevedore==5.5.0                # Carga dinámica de plugins

##############################
# 🧩 DESARROLLO INTERNO
##############################
pytokens==0.3.0                 # Manejo de tokens (autenticación, etc.)
